orders_service/requirements.txt
sqlalchemy
jose
fastapi
passlib
python-jose
psycopg2-binary
uvicorn
pydantic
pydantic[email]
Jinja2
uvicorn
kafka-python
requests
strawberry-graphql